Refactor to new aura db query runner
